% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADE2CAR.DEGs}
\alias{ADE2CAR.DEGs}
\title{DEGs of the adenoma-carcinoma transition}
\format{A  character vector}
\usage{
ADE2CAR.DEGs
}
\description{
A dataset containing 1290 DEGs of the adenoma-carcinoma transition
}
\keyword{datasets}
