% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADE2CAR.net}
\alias{ADE2CAR.net}
\title{PPI network of ADE2CAR.DEGs}
\format{A igraph object}
\usage{
ADE2CAR.net
}
\description{
A dataset containing the PPI network consisted of DEGs between colonic adenomas and carcinomas
}
\keyword{datasets}
