% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COAD.net}
\alias{COAD.net}
\title{PPI networks of the ten samples with colon adenocarcinoma}
\format{A list}
\usage{
COAD.net
}
\description{
A dataset containing PPI networks
}
\keyword{datasets}
