% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{databases.net}
\alias{databases.net}
\title{integrated PPI network}
\format{A igraph object}
\usage{
databases.net
}
\description{
A dataset containing the integrated PPI network from Biogrid, HPRD and mentha databases
}
\keyword{datasets}
