% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnetODA.R
\name{mnetODA}
\alias{mnetODA}
\title{Calculate the pair-wise ODAs among networks}
\usage{
mnetODA(netlist, mean = "arithmetic", net2edge = T)
}
\arguments{
\item{netlist}{The input networks, which should a list of graph objects}

\item{mean}{"arithmetic" or "geometric" mean to be used}

\item{net2edge}{Whether the input graph object should to transformed into a edge matrix}
}
\value{
A matrix of similarity value between 0 and 1
}
\description{
Calculate the pair-wise ODAs among networks
}
\details{
mnetODA(netlist,,mean='arithmetic',net2edge=T)
}
\examples{

 mnetODA(COAD.net)

}
