% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnetSE.R
\name{mnetSE}
\alias{mnetSE}
\title{Calculate the paired structural equivalences among networks}
\usage{
mnetSE(netlist, node, method = "cosine")
}
\arguments{
\item{netlist}{The input networks, which should a list of graph objects}

\item{node}{node for calculation of paired structural equivalences}

\item{method}{The similarity measure to be used. This must be "cosine" or "euclidean".}
}
\value{
A matrix of similarity value between 0 and 1
}
\description{
Calculate the paired structural equivalences among networks
}
\details{
mnetSE(netlist,node,method='cosine')
}
\examples{

 mnetSE(COAD.net,'APC')

}
