% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netODA.R
\name{netODA}
\alias{netODA}
\title{Calculate the orbit distribution agreement (ODA) between two networks}
\usage{
netODA(net1, net2, mean = "arithmetic", net2edge = T)
}
\arguments{
\item{net1}{The input network, which should be one of incgraph and igraph objects}

\item{net2}{The input network, which should be one of incgraph and igraph objects}

\item{mean}{"arithmetic" or "geometric" mean to be used}

\item{net2edge}{Whether the input graph object should to transformed into a edge matrix}
}
\value{
Agreement value between 0 and 1
}
\description{
Calculate the orbit distribution agreement (ODA) between two networks
}
\details{
netODA(net1,net2,mean='arithmetic',net2edge=T)
}
\examples{

 netODA(TCGA-AF-2687.net, TCGA-A6-2686.net)

}
