% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netSE.R
\name{netSE}
\alias{netSE}
\title{Calculate the structural equivalence between two networks}
\usage{
netSE(net1, net2, node, method = "cosine")
}
\arguments{
\item{net1}{The input network, which should be a graph object}

\item{net2}{The input network, which should be a graph object}

\item{node}{Gene symbol, which is shared between two networks}

\item{method}{The similarity measure to be used. This must be "cosine" or "euclidean".}
}
\value{
Similarity value between 0 and 1
}
\description{
Calculate the structural equivalence between two networks
}
\details{
netSE(net1,net2,node,method='cosine')
}
\examples{

 netSE(NOR2ADE.net,ADE2CAR.net,'EREG')

}
