% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netSEs.R
\name{netSEs}
\alias{netSEs}
\title{Calculate the structural equivalence of paired nodes}
\usage{
netSEs(net1, net2, nodes, method = "cosine")
}
\arguments{
\item{net1}{The input network, which should be a graph object}

\item{net2}{The input network, which should be a graph object}

\item{nodes}{The paired nodes between two networks}

\item{method}{The similarity measure to be used. This must be "cosine" or "euclidean".}
}
\value{
A vector of similarity value
}
\description{
Calculate the structural equivalence of paired nodes
}
\details{
netSEs(net1,net2,nodes,method='cosine')
}
\examples{

 paired<-nodesCom(TCGA-AF-2687.net, TCGA-A6-2686.net)

 netSEs(TCGA-AF-2687.net, TCGA-A6-2686.net, paired)

}
