% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodesCom.R
\name{nodesCom}
\alias{nodesCom}
\title{Get the common nodes from two networks}
\usage{
nodesCom(net1, net2)
}
\arguments{
\item{net1}{The input network, which should be a graph object}

\item{net2}{The input network, which should be a graph object}
}
\value{
A vector of gene symbols
}
\description{
Get the common nodes from two networks
}
\details{
subnet(net1,net2)
}
\examples{

 Aligned <- nodesCom(NOR2ADE.net,ADE2CAR.net)

}
