% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOR2ADE.DEGs}
\alias{NOR2ADE.DEGs}
\title{DEGs between colonic normals and adenomas}
\format{A  character vector}
\usage{
NOR2ADE.DEGs
}
\description{
A dataset containing 4567 differentially expressed genes between normals and adenomas
}
\keyword{datasets}
