% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOR2ADE.net}
\alias{NOR2ADE.net}
\title{PPI network of NOR2ADE.DEGs}
\format{A igraph object}
\usage{
NOR2ADE.net
}
\description{
A dataset containing the PPI network consisted of DEGs between colonic normals and adenomas
}
\keyword{datasets}
