% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_A6_2686.Muts}
\alias{TCGA_A6_2686.Muts}
\title{Mutated genes in the sample TCGA_A6_2686}
\format{A  list}
\usage{
TCGA_A6_2686.Muts
}
\description{
A dataset containing Mutated genes
}
\keyword{datasets}
