% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_A6_2686.net}
\alias{TCGA_A6_2686.net}
\title{PPI network of the sample TCGA_A6_2686}
\format{A igraph object}
\usage{
TCGA_A6_2686.net
}
\description{
A dataset containing a PPI network
}
\keyword{datasets}
