% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_AF_2687.Muts}
\alias{TCGA_AF_2687.Muts}
\title{Mutated genes in the sample TCGA_AF_2687}
\format{A  list}
\usage{
TCGA_AF_2687.Muts
}
\description{
A dataset containing Mutated genes
}
\keyword{datasets}
