% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_AF_2687.net}
\alias{TCGA_AF_2687.net}
\title{PPI network of the sample TCGA_AF_2687}
\format{A igraph object}
\usage{
TCGA_AF_2687.net
}
\description{
A dataset containing a PPI network
}
\keyword{datasets}
