% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netSE.R
\name{netSE}
\alias{netSE}
\title{Calculate the structural equivalence between two networks}
\usage{
netSE(net1, net2, node1, node2, method = "cosine")
}
\arguments{
\item{net1}{The input network, which should be a graph object}

\item{net2}{The input network, which should be a graph object}

\item{node1}{The paired node of network 1}

\item{node2}{The paired node of network 2}

\item{method}{The similarity measure to be used. This must be "cosine" or "euclidean".}
}
\value{
Similarity value between 0 and 1
}
\description{
Calculate the structural equivalence between two networks
}
\details{
netSE(net1,net2,node1,node2,method='cosine')
}
\examples{

 netSE(TCGA-AF-2687.net, TCGA-A6-2686.net,'NMT1', 'MED10')

}
