% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodesCom.R
\name{nodesCom}
\alias{nodesCom}
\title{Get the paired nodes which shared neighbors from two networks}
\usage{
nodesCom(net1, net2)
}
\arguments{
\item{net1}{The input network, which should be a graph object}

\item{net2}{The input network, which should be a graph object}
}
\value{
A matrix of paired nodes
}
\description{
Get the paired nodes which shared neighbors from two networks
}
\details{
nodesCom(net1,net2)
}
\examples{

 paired <- nodesCom(TCGA-AF-2687.net,TCGA-A6-2686.net)

}
